Extract isUserAuthorized.ts
